HBASE-3320  Compaction parameter minCompactSize should be configurable (change to getLong)
git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.90@1043818 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/regionserver/Store.java
@@ -197,7 +197,7 @@ public class Store implements HeapSize {
     this.majorCompactionTime = getNextMajorCompactTime();
 
     this.maxFilesToCompact = conf.getInt("hbase.hstore.compaction.max", 10);
-    this.minCompactSize = conf.getInt("hbase.hstore.compaction.min.size",
+    this.minCompactSize = conf.getLong("hbase.hstore.compaction.min.size",
         this.region.memstoreFlushSize);
     this.compactRatio = conf.getFloat("hbase.hstore.compaction.ratio", 1.2F);
